python/requirements.txt
async-timeout==4.0.3
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
importlib-resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
memory-profiler==0.61.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psutil==5.9.8
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.4
six==1.16.0
zipp==3.18.1
